AMR-WB: Enable VAD/DTX and use GSM compatible bit-rate as default.
AMR-WB is the bandwidth-efficient sibling of the HD audio codec G.722. Therefore, really leverage it features and enable voice-activation detection (VAD) so the RTP packets are transmitted discontinuously (DTX). Although this works with all versions of the OpenCORE-AMR encoder, at least version 0.1.5 is recommended (comes in Debian 11 and Ubuntu 20.04 LTS). Furthermore, to reduce the bandwidth even more, do not use the highest (AMR-WB Mode 8) but the GSM compatible bitrate as default (AMR-WB Mode 2). That still gives HD audio but works not only on UMTS (3G) but even on GSM links (2G).